Fixed some bugs, added new features
