+++ src/components/FormBuilder/FormBuilder.tsx
@@ -1,96 +1,35 @@
 import React, { useState, useEffect } from "react";
 
 import ConfigProps from "src/interfaces/config-props";
-import ValidationsProps from "src/interfaces/config-validations";
-import InputConfigsProps from "src/types/InputConfigs";
 
 import Button from "../Button/Button";
 import Input from "../Input/Input";
 import PasswordInput from "../PasswordInput/PasswordInput";
 
-import * as util from "src/util";
 import ErrorBlock from "../ErrorBlock/ErrorBlock";
+import { useInput } from "src/hooks/useInput";
 
 interface FormBuilderProps {
   config: ConfigProps[];
   setInputValues?: Function;
   aboveLink?: React.ReactNode;
+  label: string;
 }
 
 const FormBuilder: React.FC<FormBuilderProps> = ({
   config,
   setInputValues,
   aboveLink,
+  label,
 }) => {
-  const [itemProperties, setItemProperties] = useState<InputConfigsProps[]>([]);
+  const { checkValidation, handleSubmit, itemProperties, setValue } = useInput({
+    config,
+    setInputValues,
+  });
 
-  useEffect(() => {
-    const newItemProperties: InputConfigsProps[] = config.map(
-      (formElement) => ({ ...formElement, value: "", validError: "" })
-    );
-    setItemProperties(newItemProperties);
-  }, []);
-
-  const setValue = (e: React.ChangeEvent<HTMLInputElement>, index: number) => {
-    const newItemProperties = [...itemProperties];
-    newItemProperties[index].value = e.target.value;
-    setItemProperties(newItemProperties);
-  };
-
-  const getError = (
-    value: string,
-    validations: ValidationsProps[] | undefined = [],
-    index: number
-  ) => {
-    let res = "";
-    for (const validation of validations) {
-      switch (validation.validName) {
-        case "minLength":
-          if (util.checkMinLength(value, validation.validValue)) {
-            res += `${validation.validValue} is minimal number of symbols`;
-          }
-          break;
-        case "maxLength":
-          if (util.checkMaxLength(value, validation.validValue)) {
-            res += `${validation.validValue} is maximum number of symbols`;
-          }
-          break;
-        case "isEmail":
-          if (!util.checkEmail(value)) {
-            res += `Incorrect email\n`;
-          }
-          break;
-        case "isSame":
-          const item = itemProperties
-            .filter((item) => item.name === validation.validValue)
-            .filter((valueNeed) => util.checkSameValue(value, valueNeed.value));
-          if (item.length === 0) {
-            res += `Not corresponds to ${validation.validValue}\n`;
-          }
-          break;
-        default:
-          return;
-      }
-    }
-    return res;
-  };
-
-  const checkValidation = (index: number) => {
-    const newItemProperties = [...itemProperties];
-    if (newItemProperties[index].validations) {
-      newItemProperties[index].validError = getError(
-        newItemProperties[index].value,
-        newItemProperties[index].validations,
-        index
-      );
-    }
-    setItemProperties(newItemProperties);
-  };
-
-  const handleSubmit = (e: React.ChangeEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    setInputValues?.(itemProperties);
-  };
+  const isDisabled = itemProperties.some(
+    (item) => item.validError !== "" || item.value === ""
+  );
 
   return (
     <>
@@ -118,12 +57,11 @@ const FormBuilder: React.FC<FormBuilderProps> = ({
                 />
               )}
               {item.validError && <ErrorBlock>{item.validError}</ErrorBlock>}
-              
             </>
           );
         })}
         {aboveLink}
-        <Button disabled>Register</Button>
+        <Button disabled={isDisabled}>{label}</Button>
       </form>
     </>
   );
