+++ src/components/FormWrapper/style.scss
@@ -1,0 +1,25 @@
+.formwrapper {
+  &__wrapper {
+    text-align: center;
+    position: absolute;
+    left: 50%;
+    top: 50%;
+    transform: translate(-50%, -50%);
+    min-height: 600px;
+    min-width: 370px;
+    background-color: white;
+    border-radius: 15px;
+  }
+  &__title {
+    color: #3949ab;
+    font-weight: 900;
+  }
+  &__subtitle {
+    color: #3949ab;
+    font-weight: 700;
+  }
+  &__text {
+    color: #3949ab;
+    font-weight: 400;
+  }
+}
