+++ src/pages/Login/Login.tsx
@@ -3,6 +3,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 
 import FormBuilder from "src/components/FormBuilder/FormBuilder";
+import FormWrapper from "src/components/FormWrapper/FormWrapper";
 
 import config from "./config";
 
@@ -11,10 +12,8 @@ import "./style.scss";
 const Login = () => {
   return (
     <>
-      <div className="login__wrapper">
-        <h1 className="login__title">Welcome back</h1>
-        <h2 className="login__subtitle">Please log in to continue!</h2>
-
+      <FormWrapper title={'Welcome Back!'} subtitle={'Please log in to continue!'}>
+        {" "}
         <FormBuilder
           aboveLink={
             <Link className="login__toForgot" to="/forgot_password">
@@ -22,12 +21,12 @@ const Login = () => {
             </Link>
           }
           config={config}
+          label={"Login"}
         />
-
         <p className="login__toRegistration">
           Don't have an account? <Link to="/registration">Sign Up</Link>
         </p>
-      </div>
+      </FormWrapper>
     </>
   );
 };
