+++ src/pages/MainPage/MainPage.tsx
@@ -3,6 +3,9 @@ import React from "react";
 import './style.scss'
 
 const MainPage = () => {
+
+
+
   return <div></div>;
 };
