+++ src/pages/Registration/Registration.tsx
@@ -3,6 +3,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 
 import FormBuilder from "src/components/FormBuilder/FormBuilder";
+import FormWrapper from "src/components/FormWrapper/FormWrapper";
 
 import config from "./config";
 
@@ -11,15 +12,16 @@ import "./style.scss";
 const Registration = () => {
   return (
     <>
-      <div className="registration__wrapper">
-        <h1 className="registration__title">Welcome</h1>
-        <h2 className="registration__subtitle">Create your account</h2>
-        <h3 className="registration__text">Please register to continue!</h3>
-        <FormBuilder config={config} />
+      <FormWrapper
+        title={"Welcome"}
+        subtitle={"Create your Account"}
+        text={"Please register to continue!"}
+      >
+        <FormBuilder config={config} label={"Register"} />
         <p>
           Alredy have an account? <Link to="/login">Log in</Link>
         </p>
-      </div>
+      </FormWrapper>
     </>
   );
 };
